refactor: use services naming scheme
