+++ rgdps/adapters/storage.py
