+++ rgdps/api/gd/save_data.py
@@ -12,7 +12,7 @@ from rgdps.api.dependencies import authenticate_dependency
 from rgdps.common.validators import GameSaveData
 from rgdps.constants.errors import ServiceError
 from rgdps.models.user import User
-from rgdps.usecases import save_data
+from rgdps.services import save_data
 
 
 async def save_data_get(
